requirements.txt
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
